Enviando documentos de cobrança

+++ documents/documents.py
@@ -1,6 +1,6 @@
 from rest_framework.exceptions import APIException
-from documents.models import Cabecalho, Info_Adicionais, Documento_Instalacao, Documento, Documento_Pos_Venda
-from documents.serializer import CabecalhoSerializer, DocumentoInstalacaoSerializer, DocumentoPosVendasSerializer
+from documents.models import Cabecalho, Documento_Instalacao, Documento, Documento_Pos_Venda, Cobranca
+from documents.serializer import CobrancaSerializer, CabecalhoSerializer, DocumentoInstalacaoSerializer, DocumentoPosVendasSerializer
 from rest_framework.response import Response
 from products.models import Product
 from accounts.models import User
@@ -12,16 +12,13 @@ def criar_cabecalho(request):
     if cabecalho is None or not cabecalho_validado(cabecalho):
         raise APIException("O cabeçalho do documento não foi enviado ou não foi enviado adequadamente.")
     
-    if not 'informacoes_adicionais' in cabecalho or not informacoes_adicionais_validado(cabecalho['informacoes_adicionais']):
-        raise APIException("As informações adicionais do documento não foi enviado ou não foi enviado adequadamente.")
-    
-    novo_info_adicionais = Info_Adicionais.objects.create(
+    """novo_info_adicionais = Info_Adicionais.objects.create(
         distancia = cabecalho['informacoes_adicionais']['distancia'],
         horas = cabecalho['informacoes_adicionais']['horas'],
         valor_km = cabecalho['informacoes_adicionais']['valor_km'],
         valor_hora = cabecalho['informacoes_adicionais']['valor_hora'],
         total = cabecalho['informacoes_adicionais']['total']
-    )
+    )"""
     
     novo_cabecalho = Cabecalho.objects.create(
         nome=cabecalho['nome'],
@@ -34,26 +31,16 @@ def criar_cabecalho(request):
         total=cabecalho['total'],
         comissao=cabecalho['comissao'],
         email=cabecalho['email'],
-        usuario_criador=request.user,
-        info_adicionais=novo_info_adicionais
+        usuario_criador=request.user
     )
 
     return novo_cabecalho
     
 def cabecalho_validado(cabecalho):
-    campos = ['nome', 'cnpj', 'cpf', 'endereco', 'cidade', 'cep', 'telefone', 'total', 'email', 'comissao', 'informacoes_adicionais']
+    campos = ['nome', 'cnpj', 'cpf', 'endereco', 'cidade', 'cep', 'telefone', 'total', 'email', 'comissao']
 
     for campo in campos:
         if campo not in cabecalho or cabecalho[campo] == '' or cabecalho[campo] == 0:
-            return False
-
-    return True
-
-def informacoes_adicionais_validado(informacoes_adicionais):
-    campos = ['distancia', 'horas', 'valor_km', 'valor_hora', 'total']
-
-    for campo in campos:
-        if campo not in informacoes_adicionais or informacoes_adicionais[campo] == 0:
             return False
 
     return True
@@ -98,7 +85,8 @@ def criar_documentos_de_instalacao(request, cabecalho):
             servicos_executados = documento['servicos_executados'],
             testes_realizados = documento['testes_realizados'],
             total = documento['total'],
-            cabecalho = cabecalho
+            cabecalho = cabecalho,
+            comissao=documento['comissao']
         )
 
         novo_documento = Documento_Instalacao.objects.create(
@@ -136,7 +124,8 @@ def documentos_instalacao_validados(documentos):
         "total",
         "dono",
         "nota_fiscal",
-        "parceiros"
+        "parceiros",
+        "comissao"
     ]
 
     for documento_instalacao in documentos:
@@ -168,6 +157,7 @@ def criar_documentos_de_pos_vendas(request, cabecalho):
         servicos_executados = documento['servicos_executados'],
         testes_realizados = documento['testes_realizados'],
         total = documento['total'],
+        comissao=documento['comissao'],
         cabecalho = cabecalho
     )
 
@@ -191,7 +181,73 @@ def documento_pos_venda_validado(documento):
         "valor_produto",
         "servicos_executados",
         "testes_realizados",
-        "total"
+        "total",
+        "comissao"
+    ]
+
+    for campo in campos:
+        if campo not in documento or documento[campo] == '':
+            return False
+    
+    return True
+
+def criar_documento_cobranca(request, cabecalho):
+    documento = request.data.get('documento').get('cobranca')
+
+    if not documento or documento is None:
+        return None
+
+    if not documento_cobranca_validado(documento):
+        raise APIException("Documento de pós-venda está incorreto ou faltando dados.")
+    
+    produto = get_produto_por_id(documento['produto'])
+
+    novo_documento_geral = Documento.objects.create(
+        maquina = documento['maquina'],
+        numero_maquina = documento['numero_maquina'],
+        quantidade_linhas = documento['quantidade_linhas'],
+        maquina_nova =  documento['maquina_nova'],
+        faturado_revenda =  documento['faturado_revenda'],
+        produto = produto,
+        valor_produto = produto.price_setup,
+        servicos_executados = documento['servicos_executados'],
+        testes_realizados = documento['testes_realizados'],
+        total = documento['total'],
+        comissao= documento['comissao'],
+        cabecalho = cabecalho
+    )
+
+    novo_documento = Cobranca.objects.create(
+        documento = novo_documento_geral,
+        distancia = documento['distancia'],
+        horas = documento['horas'],
+        valor_km = documento['valor_km'],
+        valor_hora = documento['valor_hora'],
+        total = documento['total']
+    )
+
+    return novo_documento
+
+
+
+def documento_cobranca_validado(documento):
+    
+    campos = [  
+        "maquina",
+        "numero_maquina",
+        "quantidade_linhas",
+        "maquina_nova", 
+        "faturado_revenda",
+        "produto",
+        "valor_produto",
+        "servicos_executados",
+        "testes_realizados",
+        "total",
+        "distancia",
+        "horas",
+        "valor_km",
+        "valor_hora",
+        "comissao"
     ]
 
     for campo in campos:
@@ -213,24 +269,26 @@ def criar_documento_completo(request):
 
     resposta = {}
     
-    
     cabecalho = None
     documentos_instalacao = []
     documento_pos_venda = None
+    cobranca = None
 
     try:
         cabecalho = criar_cabecalho(request)
         documentos_instalacao = criar_documentos_de_instalacao(request, cabecalho)
         documento_pos_venda = criar_documentos_de_pos_vendas(request, cabecalho)
+        cobranca = criar_documento_cobranca(request, cabecalho)
     except APIException as erro:
         if cabecalho is not None:
-            cabecalho.info_adicionais.delete()
             cabecalho.delete()
         if len(documentos_instalacao) > 0:
             for documento_instalacao in documentos_instalacao:
                 documento_instalacao.delete()
         if documento_pos_venda is not None:
             documento_pos_venda.delete()
+        if cobranca is not None:
+            cobranca.delete()
 
         raise erro
 
@@ -246,7 +304,11 @@ def criar_documento_completo(request):
         serializer3 = DocumentoPosVendasSerializer(documento_pos_venda)
         resposta['documento_pos_venda'] = serializer3.data
     
-    if len(documentos_instalacao) == 0 and documento_pos_venda is None:
+    if cobranca is not None:
+        serializer4 = CobrancaSerializer(cobranca)
+        resposta['cobranca'] = serializer4.data
+    
+    if len(documentos_instalacao) == 0 and documento_pos_venda is None and cobranca is None:
         cabecalho.info_adicionais.delete()
         cabecalho.delete()
         raise APIException('Nenhum documento foi enviado!')
